replace using `java.util.concurrent.Future`.
This commit replaces using `java.util.concurrent.Future` into
`FutureResponse`. It provides `close()` method to close resouces
on the server which corresponded to the original request of the future.

You can use it with try-with-resources style, like as following:

```java
try (var future = client.request(something)) {
    var result = future.get();
}
```

Or you can use `await()` to shorten the above example:

```java
var result = client.request(something).await();
```

The `await()` will close resources if the request was failed.

Note that, `FutureResponse` may throw `ServerException` instead of
`ExecutionException`, that means "the request was failed in the server".